#3234 - renamed Cache -> Caching
